Moving Entities to new folder for best stucture, and creating fluent migrations of all entities.
