+++ AppServices/Services/BaseService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Domain;
+using Domain.Entities;
 
 namespace AppServices.Services
 {
