+++ AppServices/Services/JobsService.cs
@@ -1,4 +1,5 @@
 ﻿using Domain;
+using Domain.Entities;
 using System.Collections.Generic;
 using System.Linq;
