+++ Domain/Entities/Company.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-namespace Domain
+namespace Domain.Entities
+
 {
     public class Company : Entity
     {
