+++ Domain/Entities/HireType.cs
@@ -1,10 +1,11 @@
 ﻿using System;
-namespace Domain
+namespace Domain.Entities
+
 {
-    public class Category : Entity
+    public class HireType : Entity
     {
         public string Name { get; set; }
-
         public string Description { get; set; }
+        public bool PaysMoney { get; set; }
     }
 }
