+++ Domain/Entities/Permission.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-namespace Domain
+namespace Domain.Entities
+
 {
     public class Permission : Entity
     {
