+++ Domain/Entities/User.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Domain
+namespace Domain.Entities
+
 {
     public class User : Entity
     {
