+++ Domain/Framework/Constants/TableConstants.cs
@@ -3,10 +3,17 @@ namespace Domain.Framework.Constants
 {
     public class TableConstants
     {
+        public static string joelTest => "JoelTest";
+        public static string Permission => "Permission";
+        public static string Location => "Location";
+
         public static string Category => "Category";
         public static string Company => "Company";
         public static string HireType => "HireType";
         public static string Job => "Job";
         public static string Users => "Users";
+
+        public static string Role => "Role";
+        public static string Login => "Login";
     }
 }
