+++ Domain/Framework/Dto/JobPagingParameter.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Domain.Entities;
+using System;
 using System.Collections.Generic;
 
 namespace Domain.Framework.Dto
