+++ Web/ViewModels/UserProfileViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Domain;
+using Domain.Entities;
 using System.Collections.Generic;
 
 namespace Web.ViewModels
