+++ Web/Views/Home/_partials/recentJobsResultView.cshtml
@@ -1,4 +1,4 @@
-﻿@model Domain.Job
+﻿@model Domain.Entities.Job
 
 <!-- Single Large Job List -->
 <div class="job-new-list">
