+++ Web/Views/Jobs/_partials/JobResultItemView.cshtml
@@ -9,7 +9,7 @@
 
 
 }
-@model Domain.Job
+@model Domain.Entities.Job
 
 <div class="job-new-list">
     <div class="vc-thumb">
