+++ Web/Views/UserProfile/_partials/ManageJobsView.cshtml
@@ -1,4 +1,4 @@
-﻿@model List<Domain.Job>
+﻿@model List<Domain.Entities.Job>
 
 <!-- Manage jobs -->
 <div class="tab-pane" id="manage-jobs">
